+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
-# v5.14.3
+# v5.14.4
 
 Fixes:
-- Creation of a password-string resource
+- Edit mode in PCD
+    - Ensures the operator remains owner
+    - Ensures the operator in updated mode can submit the form
+    - Invert HTTP request to avoid information leak
+- Fixes resource creation, resource update and folder creation redirections